camera/requirements.txt
click==7.1.2
cs50==5.0.4
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.2.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
termcolor==1.1.0
Werkzeug==1.0.1
